+++ src/store/uiStore.ts
@@ -1,12 +1,11 @@
-import { create } from 'zustand';
-
-interface UIState {
-    isRelaxMode: boolean;
-    toggleRelaxMode: () => void;
-}
+import { UIState } from "@/types/ui";
+import { create } from "zustand";
 
 export const useUIStore = create<UIState>((set) => ({
-    isRelaxMode: false,
-    toggleRelaxMode: () =>
-        set((state) => ({ isRelaxMode: !state.isRelaxMode })),
+  isRelaxMode: false,
+  toggleRelaxMode: () => set((state) => ({ isRelaxMode: !state.isRelaxMode })),
+
+  isSidebarOpen: false,
+  toggleSidebar: () =>
+    set((state) => ({ isSidebarOpen: !state.isSidebarOpen })),
 }));
